Fix: Correctly handle SVG coordinates during drag and resize in VectorEditModal
The calculation of `deltaX` and `deltaY` in the `VectorEditModal` component was incorrect during drag and resize operations.  This resulted in inaccurate positioning and resizing of vector assets.  This commit corrects the calculation by dividing the delta values by `EDIT_MODAL_SCALE` to account for the difference between the modal's coordinate system and the underlying SVG coordinate system.  The `EDIT_MODAL_SCALE` constant was also added to the component.  The TODO comment indicates a future need to calculate this scale more precisely.