Add Z-index collision detection to ImageEditModal
This commit adds Z-index collision detection to the `ImageEditModal` component.  A new function `validateZIndex` is introduced to check for Z-index conflicts when a user modifies the Z-index of an image. The result of the validation (including potential errors or warnings) is stored in the `zIndexValidation` state variable, using a newly defined `ZIndexValidationResult` type.  The `validateZIndexNumber` function is now utilized for this validation. The Z-index is now validated upon change.